move canvas functions to codemirror package + fix id collisions
