+++ packages/widgets/index.mjs
@@ -1,2 +1,1 @@
 export * from './Claviature.jsx';
-export * from './canvas.mjs';
